Scripting cleanup
